requirements.txt
Django==1.5.1
Pillow==1.7.2
South==0.7.3
django-registration==0.7
django-taggit==0.9.3
wsgiref==0.1.2
django-cities
django-image-cropping
django-taggit
simplejson==1.9.3
twilio
django-money
django-turbolinks
py-moneyed